move expert_uid to hivemind.server
